Move `download_manager`
