+++ src/tribler/core/conftest.py
@@ -3,12 +3,9 @@ import logging
 import platform
 import sys
 import time
-from unittest.mock import MagicMock
 
 import pytest
 
-from tribler.core.components.libtorrent.download_manager.download_manager import DownloadManager
-from tribler.core.components.libtorrent.settings import LibtorrentSettings
 from tribler.core.utilities.network_utils import default_network_utils
 
 # Enable origin tracking for coroutine objects in the current thread, so when a test does not handle
@@ -47,22 +44,3 @@ def event_loop():
     loop = policy.new_event_loop()
     yield loop
     loop.close()
-
-
-@pytest.fixture
-async def download_manager(tmp_path_factory):
-    config = LibtorrentSettings()
-    config.dht = False
-    config.upnp = False
-    config.natpmp = False
-    config.lsd = False
-    download_manager = DownloadManager(
-        config=config,
-        state_dir=tmp_path_factory.mktemp('state_dir'),
-        notifier=MagicMock(),
-        peer_mid=b"0000")
-    download_manager.metadata_tmpdir = tmp_path_factory.mktemp('metadata_tmpdir')
-    download_manager.initialize()
-    yield download_manager
-
-    await download_manager.shutdown()
